fix: backend security hardening + anti-slop consolidation
Security: gRPC stream interceptors, NATS TLS enforcement, Redis ACL
support, audit logging for auth failures, WebSocket channel bounds,
expr-lang sandbox, OIDC JWKS singleflight, session invalidation.

Code quality: shared httputil.SafeClient, shared validation.WorkflowStepID,
shared maputil.CloneStringMap, parsePagination helper, rename
isAsyncFailClosed, fix error swallowing in workflow engine.

+++ core/workflow/engine_helpers_test.go
@@ -345,7 +345,7 @@ func TestHelperFunctions(t *testing.T) {
 
 	input := map[string]any{"k": map[string]any{"v": 1}}
 	clone := cloneMap(input)
-	if clone["k"].(map[string]any)["v"].(float64) != 1 {
+	if clone["k"].(map[string]any)["v"].(int) != 1 {
 		t.Fatalf("unexpected clone")
 	}
 }
